requirements.txt
Flask
Flask-SocketIO
SpeechRecognition
gTTS
groq
python-socketio
eventlet
python-dotenv